Python_MATLAB_Codes/train_inference_python/requirements.txt
tensorflow-gpu==2.5.0
imageio
tifffile
scipy==1.7.1
opencv-python
